requirements.txt
aiohttp
requests
pandas
PyJWT
cryptography
